Fix integer overflow in byte operations

+++ usecases/byte_operations/byte_operations.go
@@ -24,7 +24,7 @@ const (
 )
 
 type ByteOperations struct {
-	Position uint32
+	Position uint64
 	Buffer   []byte
 }
 
@@ -43,7 +43,7 @@ func (bo *ByteOperations) ReadUint32() uint32 {
 	return binary.LittleEndian.Uint32(bo.Buffer[bo.Position-uint32Len : bo.Position])
 }
 
-func (bo *ByteOperations) CopyBytesFromBuffer(length uint32) ([]byte, error) {
+func (bo *ByteOperations) CopyBytesFromBuffer(length uint64) ([]byte, error) {
 	out := make([]byte, length)
 	bo.Position += length
 	numCopiedBytes := copy(out, bo.Buffer[bo.Position-length:bo.Position])
@@ -69,7 +69,7 @@ func (bo *ByteOperations) WriteUint16(value uint16) {
 }
 
 func (bo *ByteOperations) CopyBytesToBuffer(copyBytes []byte) error {
-	lenCopyBytes := uint32(len(copyBytes))
+	lenCopyBytes := uint64(len(copyBytes))
 	bo.Position += lenCopyBytes
 	numCopiedBytes := copy(bo.Buffer[bo.Position-lenCopyBytes:bo.Position], copyBytes)
 	if numCopiedBytes != int(lenCopyBytes) {
@@ -78,7 +78,7 @@ func (bo *ByteOperations) CopyBytesToBuffer(copyBytes []byte) error {
 	return nil
 }
 
-func (bo *ByteOperations) MoveBufferPositionForward(length uint32) {
+func (bo *ByteOperations) MoveBufferPositionForward(length uint64) {
 	bo.Position += length
 }
